Tidying up error messaging
The `rpm-ostree db list` and `rpm-ostree db version` commands currently
exit silently if a `COMMIT` argument is not passed to them.  This change
modifies the behavior to be more like the `rpm-ostree rebase` command
when a `REFSPEC` is missing.  That is to say, an error message is
printed before the command exits.

(My familiarity with the C language is nearly nil, so I basically copied
the code used in `rpmostree-builtin-rebase.c` that handles the missing
argument.)

Signed-off-by: Micah Abbott <miabbott@redhat.com>